requirements.txt
streamlit
newspaper3k
google-generativeai
moviepy
gtts
requests
beautifulsoup4
lxml
youtube-transcript-api
Pillow
numpy
